Split out shading code into shader.rs
